Adapted tests according to the changes
